Update pointers to link capturing APIs.
For now, just link to the doc that explains the past and future plans
for link capturing. This has more context and is more understandable
than the launch handler and scope extensions explainers. When the
TODO about moving this document into the sw-launch repo is fixed, this
link should be updated again.