Fix shortname, update permanent redirects, improve xref
This update sets the shortname of the spec to `window-placement`, updates URLs
that return a permanent redirect (to save a redirect), and slightly improves
references to external specs to better use xref mechanisms and get taken into
account by Bikeshed when it builds its index and list of references.

Consideration sections would still need to be flagged as informative at some
point so that external references they contain don't get listed as normative
references but rather as informative references. Not done in this update because
it triggers a bunch of "RFC2119 words in informative section" warnings.